+++ src/main/scala/edu/ie3/simona/agent/grid/DBFSAlgorithm.scala
@@ -22,9 +22,9 @@ import edu.ie3.simona.agent.grid.GridAgentData.{
 }
 import edu.ie3.simona.agent.grid.GridAgentMessages.Responses.ExchangeVoltage
 import edu.ie3.simona.agent.grid.GridAgentMessages._
-import edu.ie3.simona.agent.participant.ParticipantAgent.{
-  FinishParticipantSimulation,
-  ParticipantMessage,
+import edu.ie3.simona.agent.participant2.ParticipantAgent
+import edu.ie3.simona.agent.participant2.ParticipantAgent.{
+  GridSimulationFinished,
   RequestAssetPowerMessage,
 }
 import edu.ie3.simona.event.RuntimeEvent.PowerFlowFailed
@@ -35,14 +35,12 @@ import edu.ie3.simona.ontology.messages.SchedulerMessage.Completion
 import edu.ie3.util.scala.quantities.DefaultQuantities._
 import edu.ie3.util.scala.quantities.SquantsUtils.RichElectricPotential
 import org.apache.pekko.actor.typed.scaladsl.AskPattern._
-import org.apache.pekko.actor.typed.scaladsl.adapter.TypedActorRefOps
 import org.apache.pekko.actor.typed.scaladsl.{
   ActorContext,
   Behaviors,
   StashBuffer,
 }
-import org.apache.pekko.actor.typed.{ActorRef, Behavior, Scheduler}
-import org.apache.pekko.pattern.ask
+import org.apache.pekko.actor.typed.{ActorRef, ActorSystem, Behavior, Scheduler}
 import org.apache.pekko.util.{Timeout => PekkoTimeout}
 import squants.Each
 
@@ -348,8 +346,8 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
                             // To model the exchanged power from the superior grid's point of view, -1 has to be multiplied.
                             // (Inferior grid is a feed in facility to superior grid, which is negative then). Analogously for load case.
                             (
-                              refSystem.pInSi(pInPu) * (-1),
-                              refSystem.qInSi(qInPu) * (-1),
+                              refSystem.pInSi(pInPu) * -1,
+                              refSystem.qInSi(qInPu) * -1,
                             )
                           case _ =>
                             /* TODO: As long as there are no multiple slack nodes, provide "real" power only for the slack node */
@@ -437,6 +435,8 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
               FinishGridSimulationTrigger(currentTick),
               gridAgentBaseData: GridAgentBaseData,
             ) =>
+          val nextTick = currentTick + constantData.resolution
+
           // inform my child grids about the end of this grid simulation
           gridAgentBaseData.inferiorGridGates
             .map {
@@ -451,7 +451,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
           gridAgentBaseData.gridEnv.nodeToAssetAgents.foreach {
             case (_, actors) =>
               actors.foreach { actor =>
-                actor ! FinishParticipantSimulation(currentTick)
+                actor ! GridSimulationFinished(currentTick, nextTick)
               }
           }
 
@@ -1086,14 +1086,16 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
   private def askForAssetPowers(
       currentTick: Long,
       sweepValueStore: Option[SweepValueStore],
-      nodeToAssetAgents: Map[UUID, Set[ActorRef[ParticipantMessage]]],
+      nodeToAssetAgents: Map[UUID, Set[ActorRef[ParticipantAgent.Request]]],
       refSystem: RefSystem,
       askTimeout: Duration,
   )(implicit
       ctx: ActorContext[GridAgent.Request]
   ): Boolean = {
-    implicit val timeout: PekkoTimeout = PekkoTimeout.create(askTimeout)
     implicit val ec: ExecutionContext = ctx.executionContext
+
+    implicit val timeout: PekkoTimeout = PekkoTimeout.create(askTimeout)
+    implicit val system: ActorSystem[_] = ctx.system
 
     ctx.log.debug(s"asking assets for power values: {}", nodeToAssetAgents)
 
@@ -1128,16 +1130,21 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
                     )
                 }
 
-              (assetAgent.toClassic ? RequestAssetPowerMessage(
-                currentTick,
-                eInPu,
-                fInPU,
-              )).map {
-                case providedPowerValuesMessage: AssetPowerChangedMessage =>
-                  (assetAgent, providedPowerValuesMessage)
-                case assetPowerUnchangedMessage: AssetPowerUnchangedMessage =>
-                  (assetAgent, assetPowerUnchangedMessage)
-              }
+              assetAgent
+                .ask(replyTo =>
+                  RequestAssetPowerMessage(
+                    currentTick,
+                    eInPu,
+                    fInPU,
+                    replyTo,
+                  )
+                )
+                .map {
+                  case providedPowerValuesMessage: AssetPowerChangedMessage =>
+                    (assetAgent, providedPowerValuesMessage)
+                  case assetPowerUnchangedMessage: AssetPowerUnchangedMessage =>
+                    (assetAgent, assetPowerUnchangedMessage)
+                }
             })
           }.toVector
         )