+++ src/main/scala/edu/ie3/simona/agent/grid/GridAgent.scala
@@ -22,7 +22,8 @@ import edu.ie3.simona.agent.grid.congestion.{
   CongestionManagementParams,
   DCMAlgorithm,
 }
-import edu.ie3.simona.agent.participant.ParticipantAgent.ParticipantMessage
+import edu.ie3.simona.agent.grid.GridAgentMessages._
+import edu.ie3.simona.agent.participant2.ParticipantAgent
 import edu.ie3.simona.config.SimonaConfig
 import edu.ie3.simona.event.ResultEvent
 import edu.ie3.simona.event.ResultEvent.PowerFlowResultEvent
@@ -182,7 +183,8 @@ object GridAgent extends DBFSAlgorithm with DCMAlgorithm {
         )
 
       /* Reassure, that there are also calculation models for the given uuids */
-      val nodeToAssetAgentsMap: Map[UUID, Set[ActorRef[ParticipantMessage]]] =
+      val nodeToAssetAgentsMap
+          : Map[UUID, Set[ActorRef[ParticipantAgent.Request]]] =
         gridAgentController
           .buildSystemParticipants(subGridContainer, thermalGridsByBusId)
           .map { case (uuid: UUID, actorSet) =>