+++ src/main/scala/edu/ie3/simona/config/SimonaConfig.scala
@@ -92,52 +92,11 @@ object SimonaConfig {
       val notifier: String,
       val simulationResult: Boolean,
   )
-
-  sealed abstract class BaseRuntimeConfig(
-      val calculateMissingReactivePowerWithModel: Boolean = false,
-      val scaling: Double = 1.0,
-      val uuids: List[String] = List.empty,
-  ) extends Serializable
-
   sealed abstract class CsvParams(
       val csvSep: String,
       val directoryPath: String,
       val isHierarchic: Boolean,
   )
-
-  final case class EmRuntimeConfig(
-      override val calculateMissingReactivePowerWithModel: Boolean = false,
-      override val scaling: Double = 1.0,
-      override val uuids: List[String] = List.empty,
-      aggregateFlex: String = "SELF_OPT_EXCL_REG",
-      curtailRegenerative: Boolean = false,
-  ) extends BaseRuntimeConfig(
-        calculateMissingReactivePowerWithModel,
-        scaling,
-        uuids,
-      )
-
-  final case class EvcsRuntimeConfig(
-      override val calculateMissingReactivePowerWithModel: Boolean = false,
-      override val scaling: Double = 1.0,
-      override val uuids: List[String] = List.empty,
-      chargingStrategy: String = "maxPower",
-      lowestEvSoc: Double = 0.2,
-  ) extends BaseRuntimeConfig(
-        calculateMissingReactivePowerWithModel,
-        scaling,
-        uuids,
-      )
-
-  final case class FixedFeedInRuntimeConfig(
-      override val calculateMissingReactivePowerWithModel: Boolean = false,
-      override val scaling: Double = 1.0,
-      override val uuids: List[String] = List.empty,
-  ) extends BaseRuntimeConfig(
-        calculateMissingReactivePowerWithModel,
-        scaling,
-        uuids,
-      )
 
   final case class GridOutputConfig(
       congestions: Boolean = false,
@@ -149,34 +108,12 @@ object SimonaConfig {
       transformers3w: Boolean = false,
   )
 
-  final case class HpRuntimeConfig(
-      override val calculateMissingReactivePowerWithModel: Boolean = false,
-      override val scaling: Double = 1.0,
-      override val uuids: List[String] = List.empty,
-  ) extends BaseRuntimeConfig(
-        calculateMissingReactivePowerWithModel,
-        scaling,
-        uuids,
-      )
-
   sealed abstract class KafkaParams(
       val bootstrapServers: String,
       val linger: Int,
       val runId: String,
       val schemaRegistryUrl: String,
   )
-
-  final case class LoadRuntimeConfig(
-      override val calculateMissingReactivePowerWithModel: Boolean = false,
-      override val scaling: Double = 1.0,
-      override val uuids: List[String] = List.empty,
-      modelBehaviour: String,
-      reference: String,
-  ) extends BaseRuntimeConfig(
-        calculateMissingReactivePowerWithModel,
-        scaling,
-        uuids,
-      )
 
   final case class ParticipantBaseOutputConfig(
       override val notifier: String,
@@ -191,16 +128,6 @@ object SimonaConfig {
       override val isHierarchic: Boolean,
       timePattern: String = "yyyy-MM-dd'T'HH:mm:ss[.S[S][S]]X",
   ) extends CsvParams(csvSep, directoryPath, isHierarchic)
-
-  final case class PvRuntimeConfig(
-      override val calculateMissingReactivePowerWithModel: Boolean = false,
-      override val scaling: Double = 1.0,
-      override val uuids: List[String] = List.empty,
-  ) extends BaseRuntimeConfig(
-        calculateMissingReactivePowerWithModel,
-        scaling,
-        uuids,
-      )
 
   sealed trait GridConfigParams {
     val gridIds: Option[List[String]]
@@ -235,18 +162,6 @@ object SimonaConfig {
       override val simulationResult: Boolean,
   ) extends BaseOutputConfig(notifier, simulationResult)
 
-  final case class StorageRuntimeConfig(
-      override val calculateMissingReactivePowerWithModel: Boolean = false,
-      override val scaling: Double = 1.0,
-      override val uuids: List[String] = List.empty,
-      initialSoc: Double = 0d,
-      targetSoc: Option[Double],
-  ) extends BaseRuntimeConfig(
-        calculateMissingReactivePowerWithModel,
-        scaling,
-        uuids,
-      )
-
   final case class TransformerControlGroup(
       measurements: List[String] = List.empty,
       transformers: List[String] = List.empty,
@@ -266,16 +181,6 @@ object SimonaConfig {
       override val voltLvls: Option[List[VoltLvlConfig]] = None,
   ) extends GridConfigParams
 
-  final case class WecRuntimeConfig(
-      override val calculateMissingReactivePowerWithModel: Boolean = false,
-      override val scaling: Double = 1.0,
-      override val uuids: List[String] = List.empty,
-  ) extends BaseRuntimeConfig(
-        calculateMissingReactivePowerWithModel,
-        scaling,
-        uuids,
-      )
-
   final case class Simona(
       congestionManagement: Simona.CongestionManagement =
         Simona.CongestionManagement(),
@@ -285,7 +190,7 @@ object SimonaConfig {
       input: Simona.Input,
       output: Simona.Output,
       powerflow: Simona.Powerflow,
-      runtime: Simona.Runtime,
+      runtime: RuntimeConfig,
       simulationName: String,
       time: Simona.Time = Simona.Time(),
   )
@@ -494,72 +399,6 @@ object SimonaConfig {
           epsilon: List[Double] = List.empty,
           iterations: Int,
       )
-    }
-
-    final case class Runtime(
-        listener: Runtime.Listener = Runtime.Listener(),
-        participant: Runtime.Participant,
-        selected_subgrids: Option[List[Int]] = None,
-        selected_volt_lvls: Option[List[VoltLvlConfig]] = None,
-    )
-    object Runtime {
-      final case class Listener(
-          eventsToProcess: Option[List[String]] = None,
-          kafka: Option[RuntimeKafkaParams] = None,
-      )
-
-      final case class Participant(
-          em: Participant.Em,
-          evcs: Participant.Evcs,
-          fixedFeedIn: Participant.FixedFeedIn,
-          hp: Participant.Hp,
-          load: Participant.Load,
-          pv: Participant.Pv,
-          requestVoltageDeviationThreshold: Double = 1e-14,
-          storage: Participant.Storage,
-          wec: Participant.Wec,
-      )
-      object Participant {
-        final case class Em(
-            defaultConfig: EmRuntimeConfig,
-            individualConfigs: List[EmRuntimeConfig] = List.empty,
-        )
-
-        final case class Evcs(
-            defaultConfig: EvcsRuntimeConfig,
-            individualConfigs: List[EvcsRuntimeConfig] = List.empty,
-        )
-
-        final case class FixedFeedIn(
-            defaultConfig: FixedFeedInRuntimeConfig,
-            individualConfigs: List[FixedFeedInRuntimeConfig] = List.empty,
-        )
-
-        final case class Hp(
-            defaultConfig: HpRuntimeConfig,
-            individualConfigs: List[HpRuntimeConfig] = List.empty,
-        )
-
-        final case class Load(
-            defaultConfig: LoadRuntimeConfig,
-            individualConfigs: List[LoadRuntimeConfig] = List.empty,
-        )
-
-        final case class Pv(
-            defaultConfig: PvRuntimeConfig,
-            individualConfigs: List[PvRuntimeConfig] = List.empty,
-        )
-
-        final case class Storage(
-            defaultConfig: StorageRuntimeConfig,
-            individualConfigs: List[StorageRuntimeConfig] = List.empty,
-        )
-
-        final case class Wec(
-            defaultConfig: WecRuntimeConfig,
-            individualConfigs: List[WecRuntimeConfig] = List.empty,
-        )
-      }
     }
 
     final case class Time(