+++ src/test/scala/edu/ie3/simona/config/ConfigFailFastSpec.scala
@@ -7,6 +7,7 @@
 package edu.ie3.simona.config
 
 import com.typesafe.config.ConfigFactory
+import edu.ie3.simona.config.RuntimeConfig.StorageRuntimeConfig
 import edu.ie3.simona.config.SimonaConfig.Simona.Input.Weather.Datasource
 import edu.ie3.simona.config.SimonaConfig.Simona.Input.Weather.Datasource.{
   CoordinateSource,
@@ -38,7 +39,7 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
         "let valid input pass" in {
           noException shouldBe thrownBy {
             ConfigFailFast invokePrivate checkTimeConfig(
-              new Time(
+              Time(
                 "2020-06-18T13:41:00Z",
                 None,
                 "2020-05-18T13:41:00Z",
@@ -50,7 +51,7 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
         "identify invalid date or time configuration" in {
           intercept[InvalidConfigParameterException] {
             ConfigFailFast invokePrivate checkTimeConfig(
-              new Time(
+              Time(
                 "2020-06-18T13:41:00Z",
                 None,
                 "2020-07-18T13:41:00Z",
@@ -94,7 +95,7 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
             ConfigFailFast invokePrivate checkPowerFlowResolutionConfiguration(
               new Powerflow(
                 10,
-                new Newtonraphson(
+                Newtonraphson(
                   List(10, 30),
                   100,
                 ),
@@ -111,7 +112,7 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
             ConfigFailFast invokePrivate checkPowerFlowResolutionConfiguration(
               new Powerflow(
                 10,
-                new Newtonraphson(
+                Newtonraphson(
                   List(10, 30),
                   100,
                 ),
@@ -1329,16 +1330,16 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
 
       "throw exception if default initial SOC is negative" in {
 
-        val defaultConfig: SimonaConfig.StorageRuntimeConfig =
-          SimonaConfig.StorageRuntimeConfig(
+        val defaultConfig: StorageRuntimeConfig =
+          StorageRuntimeConfig(
             calculateMissingReactivePowerWithModel = false,
             1.0,
             List(java.util.UUID.randomUUID().toString),
             -0.5,
             Some(0.8),
           )
-        val storageConfig = SimonaConfig.Simona.Runtime.Participant
-          .Storage(defaultConfig, List.empty)
+        val storageConfig =
+          RuntimeConfig.ParticipantRuntimeConfigs(defaultConfig, List.empty)
 
         intercept[RuntimeException] {
           ConfigFailFast invokePrivate checkStorageConfigs(storageConfig)
@@ -1346,16 +1347,16 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
       }
 
       "throw exception if default target SOC is negative" in {
-        val defaultConfig: SimonaConfig.StorageRuntimeConfig =
-          SimonaConfig.StorageRuntimeConfig(
+        val defaultConfig: StorageRuntimeConfig =
+          StorageRuntimeConfig(
             calculateMissingReactivePowerWithModel = false,
             1.0,
             List(java.util.UUID.randomUUID().toString),
             0.5,
             Some(-0.8),
           )
-        val storageConfig = SimonaConfig.Simona.Runtime.Participant
-          .Storage(defaultConfig, List.empty)
+        val storageConfig =
+          RuntimeConfig.ParticipantRuntimeConfigs(defaultConfig, List.empty)
 
         intercept[RuntimeException] {
           ConfigFailFast invokePrivate checkStorageConfigs(storageConfig)
@@ -1364,16 +1365,16 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
 
       "throw exception if individual initial SOC is negative" in {
         val uuid = java.util.UUID.randomUUID().toString
-        val defaultConfig: SimonaConfig.StorageRuntimeConfig =
-          SimonaConfig.StorageRuntimeConfig(
+        val defaultConfig: StorageRuntimeConfig =
+          StorageRuntimeConfig(
             calculateMissingReactivePowerWithModel = false,
             1.0,
             List(java.util.UUID.randomUUID().toString),
             0.5,
             Some(0.8),
           )
-        val individualConfig: List[SimonaConfig.StorageRuntimeConfig] = List(
-          SimonaConfig.StorageRuntimeConfig(
+        val individualConfig: List[StorageRuntimeConfig] = List(
+          StorageRuntimeConfig(
             calculateMissingReactivePowerWithModel = false,
             1.0,
             List(uuid),
@@ -1381,8 +1382,10 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
             Some(0.8),
           )
         )
-        val storageConfig = SimonaConfig.Simona.Runtime.Participant
-          .Storage(defaultConfig, individualConfig)
+        val storageConfig = RuntimeConfig.ParticipantRuntimeConfigs(
+          defaultConfig,
+          individualConfig,
+        )
 
         intercept[RuntimeException] {
           ConfigFailFast invokePrivate checkStorageConfigs(storageConfig)
@@ -1391,16 +1394,16 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
 
       "throw exception if individual target SOC is negative" in {
         val uuid = java.util.UUID.randomUUID().toString
-        val defaultConfig: SimonaConfig.StorageRuntimeConfig =
-          SimonaConfig.StorageRuntimeConfig(
+        val defaultConfig: StorageRuntimeConfig =
+          StorageRuntimeConfig(
             calculateMissingReactivePowerWithModel = false,
             1.0,
             List(java.util.UUID.randomUUID().toString),
             0.5,
             Some(0.8),
           )
-        val individualConfig: List[SimonaConfig.StorageRuntimeConfig] = List(
-          SimonaConfig.StorageRuntimeConfig(
+        val individualConfig: List[StorageRuntimeConfig] = List(
+          StorageRuntimeConfig(
             calculateMissingReactivePowerWithModel = false,
             1.0,
             List(uuid),
@@ -1408,8 +1411,10 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
             Some(-0.8),
           )
         )
-        val storageConfig = SimonaConfig.Simona.Runtime.Participant
-          .Storage(defaultConfig, individualConfig)
+        val storageConfig = RuntimeConfig.ParticipantRuntimeConfigs(
+          defaultConfig,
+          individualConfig,
+        )
 
         intercept[RuntimeException] {
           ConfigFailFast invokePrivate checkStorageConfigs(storageConfig)
@@ -1417,16 +1422,16 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
       }
 
       "not throw exception if all parameters are in parameter range" in {
-        val defaultConfig: SimonaConfig.StorageRuntimeConfig =
-          SimonaConfig.StorageRuntimeConfig(
+        val defaultConfig: StorageRuntimeConfig =
+          StorageRuntimeConfig(
             calculateMissingReactivePowerWithModel = false,
             1.0,
             List(java.util.UUID.randomUUID().toString),
             0.5,
             Some(0.8),
           )
-        val individualConfig: List[SimonaConfig.StorageRuntimeConfig] = List(
-          SimonaConfig.StorageRuntimeConfig(
+        val individualConfig: List[StorageRuntimeConfig] = List(
+          StorageRuntimeConfig(
             calculateMissingReactivePowerWithModel = false,
             1.0,
             List(java.util.UUID.randomUUID().toString),
@@ -1434,8 +1439,10 @@ class ConfigFailFastSpec extends UnitSpec with ConfigTestData {
             Some(0.8),
           )
         )
-        val storageConfig = SimonaConfig.Simona.Runtime.Participant
-          .Storage(defaultConfig, individualConfig)
+        val storageConfig = RuntimeConfig.ParticipantRuntimeConfigs(
+          defaultConfig,
+          individualConfig,
+        )
 
         noException should be thrownBy {
           ConfigFailFast invokePrivate checkStorageConfigs(storageConfig)